# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ReportRequest {
  /**
     * @minLength 3
     * @maxLength 100
     */
  title: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  delay_seconds?: number;
  simulate_failure?: boolean;
}

export type ReportJobStatus = typeof ReportJobStatus[keyof typeof ReportJobStatus];


export const ReportJobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ReportJob {
  id: string;
  title: string;
  status: ReportJobStatus;
  created_at: string;
  updated_at: string;
  /** @nullable */
  started_at?: string | null;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  artifact_reference?: string | null;
  /** @nullable */
  error_message?: string | null;
  source_record_count: number;
}

export type SourceSummaryCategoryTotals = {[key: string]: number};

export interface SourceSummary {
  record_count: number;
  total_amount: number;
  average_amount: number;
  category_totals: SourceSummaryCategoryTotals;
}