# lib/api-zod/src/generated/types/reportJob.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportJobStatus } from './reportJobStatus';

export interface ReportJob {
  id: string;
  title: string;
  status: ReportJobStatus;
  created_at: Date;
  updated_at: Date;
  /** @nullable */
  started_at?: Date | null;
  /** @nullable */
  completed_at?: Date | null;
  /** @nullable */
  artifact_reference?: string | null;
  /** @nullable */
  error_message?: string | null;
  source_record_count: number;
}