# lib/api-zod/src/generated/types/reportRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportRequest {
  /**
     * @minLength 3
     * @maxLength 100
     */
  title: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  delay_seconds?: number;
  simulate_failure?: boolean;
}