# lib/api-zod/src/generated/types/sourceSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SourceSummaryCategoryTotals } from './sourceSummaryCategoryTotals';

export interface SourceSummary {
  record_count: number;
  total_amount: number;
  average_amount: number;
  category_totals: SourceSummaryCategoryTotals;
}